services/explainer/requirements.txt
# FairOps Explainer — Dependencies
# Pinned versions from AGENT.md Section 20

fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
httpx==0.27.0
redis==5.0.1

# Explainability
shap==0.44.0
dice-ml==0.11
reportlab==4.1.0

# ML
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0

# Gemini
google-generativeai==0.5.0

# GCP
google-cloud-bigquery==3.17.0
google-cloud-logging==3.9.0
google-cloud-secret-manager==2.18.0

opentelemetry-sdk==1.22.0
